rename utils -> records_from_urls
